Several adddon installation bug fixes
Singularity was not properly handling install of addons that did not
have a 'release' branch file. Added additional checks to fallback to
beta and alpha files if no release files exist.

Singularity was also not displaying the install addon button on the
details page due an issue with checking the wrong parameter.